+++ src/main/java/com/energy/service/ClienteService.java
@@ -1,5 +1,6 @@
 package com.energy.service;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Optional;
 import java.util.regex.Pattern;
@@ -191,7 +192,7 @@ public class ClienteService {
 	}
 
 
-	public Page<Cliente> findByFatturatoAnnuale(Pageable page, Double fatturato) {
+	public Page<Cliente> findByFatturatoAnnuale(Pageable page, BigDecimal fatturato) {
 		return clienteRepository.findByFatturatoAnnuale(page, fatturato);
 	}
